refactor(web): refactor-search-component
